requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
feedparser==6.0.12
groq==0.36.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
postgrest==2.24.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
python-telegram-bot==22.5
realtime==2.24.0
requests==2.32.5
requests-oauthlib==2.0.0
schedule==1.2.2
sgmllib3k==1.0.0
sniffio==1.3.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
flask
